+++ libs/components/src/lib/OrderFilterConsole.tsx
@@ -14,7 +14,7 @@ import {
   Text,
   useBreakpoint,
 } from '@chakra-ui/react';
-import { fmOrderStatuses } from '@project-lc/shared-types';
+import { fmOrderStatuses, OrderFilterFormType } from '@project-lc/shared-types';
 import { useFmOrderStore } from '@project-lc/stores';
 import moment from 'moment';
 import 'moment/locale/ko';
@@ -23,14 +23,6 @@ import { DateRangePicker } from 'react-dates';
 import 'react-dates/initialize';
 import 'react-dates/lib/css/_datepicker.css';
 import { Controller, SubmitHandler, useForm } from 'react-hook-form';
-
-export interface OrderFilterFormType {
-  search: string;
-  searchDateType: '주문일' | '입금일';
-  searchStartDate: string | null; // this way is not supported https://github.com/react-hook-form/react-hook-form/issues/4704
-  searchEndDate: string | null; // this way is not supported https://github.com/react-hook-form/react-hook-form/issues/4704
-  searchStatuses: string[];
-}
 
 export function OrderFilterConsole(): JSX.Element {
   const xSize = useBreakpoint();
